refactor done. it compiles. work missing
