+++ linux/net/rina/rmt.h
@@ -49,6 +49,54 @@ struct rmt;
 
 /* NOTE: There's one RMT for each IPC Process */
 
+/* Plugin support */
+
+#define RMT_PS_HASHSIZE 7
+
+enum flow_state {
+        N1_PORT_STATE_ENABLED,
+        N1_PORT_STATE_DISABLED,
+        N1_PORT_STATE_BUSY
+};
+
+struct rmt_n1_port {
+        spinlock_t             lock;
+        port_id_t              port_id;
+        struct ipcp_instance * n1_ipcp;
+        struct hlist_node      hlist;
+        enum flow_state        state;
+        atomic_t               n_sdus;
+};
+
+/* The key in this struct is used to filter by cep_ids, qos_id, address... */
+struct rmt_kqueue {
+        struct rfifo *    queue;
+        unsigned int      key;
+        unsigned int      max_q;
+        struct hlist_node hlist;
+};
+
+struct rmt_qgroup {
+        port_id_t         pid;
+        struct hlist_node hlist;
+        DECLARE_HASHTABLE(queues, RMT_PS_HASHSIZE);
+};
+
+struct rmt_queue_set {
+        spinlock_t lock;
+        DECLARE_HASHTABLE(qgroups, RMT_PS_HASHSIZE);
+};
+
+struct rmt_kqueue *     rmt_kqueue_create(unsigned int key);
+int                     rmt_kqueue_destroy(struct rmt_kqueue * q);
+struct rmt_qgroup *     rmt_qgroup_create(void);
+int                     rmt_qgroup_destroy(struct rmt_qgroup* g);
+struct rmt_kqueue *     rmt_qgroup_find(struct rmt_qgroup * g,
+                                        unsigned int        key);
+struct rmt_queue_set *  rmt_queue_set_create(void);
+int                     rmt_queue_set_destroy(struct rmt_queue_set * qs);
+struct rmt_qgroup *     rmt_queue_set_find(struct rmt_queue_set * qs,
+                                           port_id_t              pid);
 struct rmt * rmt_create(struct ipcp_instance *  parent,
                         struct kfa *            kfa,
                         struct efcp_container * efcpc);
@@ -86,7 +134,9 @@ int          rmt_send(struct rmt * instance,
 int          rmt_send_port_id(struct rmt *  instance,
                               port_id_t     id,
                               struct pdu *  pdu);
-
+int          rmt_n1_port_write(struct rmt *         rmt,
+                               struct rmt_n1_port * n1_port,
+                               struct pdu *         pdu);
 int          rmt_receive(struct rmt * instance,
                          struct sdu * sdu,
                          port_id_t    from);
@@ -106,51 +156,4 @@ int          rmt_set_policy_set_param(struct rmt * rmt,
 
 struct rmt * rmt_from_component(struct rina_component * component);
 
-/* Plugin support */
-
-#define RMT_PS_HASHSIZE 7
-
-enum flow_state {
-        N1_PORT_STATE_ENABLED,
-        N1_PORT_STATE_DISABLED,
-        N1_PORT_STATE_BUSY
-};
-
-struct rmt_n1_port {
-        spinlock_t             lock;
-        port_id_t              port_id;
-        struct ipcp_instance * n1_ipcp;
-        struct hlist_node      hlist;
-        enum flow_state        state;
-        atomic_t               n_sdus;
-};
-
-/* The key in this struct is used to filter by cep_ids, qos_id, address... */
-struct rmt_kqueue {
-        struct rfifo *    queue;
-        unsigned int      key;
-        struct hlist_node hlist;
-};
-
-struct rmt_qgroup {
-        port_id_t         pid;
-        struct hlist_node hlist;
-        DECLARE_HASHTABLE(queues, RMT_PS_HASHSIZE);
-};
-
-struct rmt_queue_set {
-        spinlock_t lock;
-        DECLARE_HASHTABLE(qgroups, RMT_PS_HASHSIZE);
-};
-
-struct rmt_kqueue *     rmt_kqueue_create(unsigned int key);
-int                     rmt_kqueue_destroy(struct rmt_kqueue * q);
-struct rmt_qgroup *     rmt_qgroup_create(void);
-int                     rmt_qgroup_destroy(struct rmt_qgroup* g);
-struct rmt_kqueue *     rmt_qgroup_find(struct rmt_qgroup * g,
-                                        unsigned int        key);
-struct rmt_queue_set *  rmt_queue_set_create(void);
-int                     rmt_queue_set_destroy(struct rmt_queue_set * qs);
-struct rmt_qgroup *     rmt_queue_set_find(struct rmt_queue_set * qs,
-                                           port_id_t              pid);
 #endif
